+++ CHANGES.rst
@@ -45,6 +45,7 @@ Release History
 - Added ``nengo_dl.configure_settings(use_loop=False)`` option to avoid building the
   simulation inside a symbolic TensorFlow loop. This may improve simulation speed,
   but the simulation can only run for exactly ``unroll_simulation`` timesteps.
+- NengoDL now requires ``jinja2`` (used to template some of the docstrings).
 
 **Changed**